+++ gxf_extensions/opengl/opengl_renderer.cpp
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: Copyright (c) 2022 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-FileCopyrightText: Copyright (c) 2022-2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -253,7 +253,7 @@ GLenum OpenGLRenderer::getTextureFormat(gxf::VideoFormat format) {
     case gxf::VideoFormat::GXF_VIDEO_FORMAT_RGBA:
       return GL_RGBA;
     default:
-      GXF_LOG_WARNING("Unsupported video format: %d", format);
+      GXF_LOG_WARNING("Unsupported video format: %ld", (int64_t)format);
       return GL_NONE;
   }
 }
